Login controller,loginMenu and restructuring of code
i created a login menu and a login controller which i put into the main funciton
when run it will call the login menu and ask the user to login. after login it will create a new object according to the user.

there is a bug now tho idk why my patientmenu loop is runnign forever